Внесены изменения после ревью

+++ src/test/java/OrderTest/CreateOrderTest.java
@@ -13,6 +13,7 @@ import org.junit.jupiter.api.Test;
 
 import static model.user.UserCreds.credsFrom;
 
+import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 
@@ -29,8 +30,8 @@ public class CreateOrderTest {
 
 
         this.user=UserGenerator.createRandom();
-        Response responseCreate = (Response) apiUser.createNewUserStep(this.user);
-        assertEquals(200, responseCreate.statusCode(), "Не удалось создать пользователя через API");
+        Response createResponse = apiUser.createNewUserStep(this.user);
+        assertEquals(200, createResponse.statusCode(), "Не удалось создать пользователя через API");
 
         Response loginResponse = apiUser.loginUserStep(credsFrom(user));
         accessToken = loginResponse.as(UserLoginResponse.class).getAccessToken();
@@ -38,41 +39,18 @@ public class CreateOrderTest {
 
         Order order = new Order("61c0c5a71d1f82001bdaaa70");
         Response response = orderApi.createOrderWithAuthStep(accessToken, order);
-        assertEquals(200, response.statusCode());
+
+        response.then()
+                .assertThat()
+                .statusCode(200)
+                .body("order", notNullValue());
         System.out.println(response.body().asString());
     }
 
-    @Test
-    @DisplayName("Создание заказа без авторизации")
-    public void createOrderNotAuthTest() {
-
-        Order order = new Order("61c0c5a71d1f82001bdaaa70");
-        Response response = orderApi.createOrderNotAuthStep(order);
-        assertEquals(401, response.statusCode());
-        System.out.println(response.body().asString());
-    }
-
-    @Test
-    @DisplayName("Создание заказа без ингредиентов")
-    public void createOrderNotIngridientsTest() {
-
-        Order order = new Order(null);
-        Response response = orderApi.createOrderNotAuthStep(order);
-        assertEquals(400, response.statusCode());
-        System.out.println(response.body().asString());
-    }
-
-    @Test
-    @DisplayName("Создание заказа с неверным хэшем ингредиентов")
-    public void createOrderNotValidIngridientsTest() {
-
-        Order order = new Order("hfjfjgfjgjg");
-        Response response = orderApi.createOrderNotAuthStep(order);
-        assertEquals(500, response.statusCode());
-        System.out.println(response.body().asString());
-    }
     @AfterEach
     public void tearDown() {
-        apiUser.deleteUserStep(accessToken);
+        if (accessToken != null) {
+            apiUser.deleteUserStep(accessToken);
+        }
     }
 }
