+++ src/test/java/UserTest/LoginUserTest.java
@@ -8,40 +8,53 @@ import org.junit.jupiter.api.Test;
 
 
 import static model.user.UserCreds.credsFrom;
+import static org.hamcrest.CoreMatchers.notNullValue;
+import static org.hamcrest.Matchers.equalTo;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 
 public class LoginUserTest {
 
     private ApiUser apiUser = new ApiUser();
-    private String accessToken;
     private User user;
+    private String accessToken;
 
     @Test
-    @DisplayName ("Авторизация с существующим логином")
+    @DisplayName("Авторизация с существующим логином")
     public void existingUserLoginTest() {
-        this.user= UserGenerator.createRandom();
-        Response responseCreate = (Response) apiUser.createNewUserStep(this.user);
+
+        this.user = UserGenerator.createRandom();
+        Response responseCreate = apiUser.createNewUserStep(this.user);
         assertEquals(200, responseCreate.statusCode(), "Не удалось создать пользователя через API");
 
-    UserCreds creds = credsFrom(user);
-    Response loginResponse = apiUser.loginUserStep(creds);
-    String accessToken = loginResponse.as(UserLoginResponse.class).getAccessToken();
-    assertEquals(200, loginResponse.statusCode());
-    assertNotNull(accessToken, "Полученный токен пуст");
-}
+        UserCreds creds = credsFrom(user);
+        Response loginResponse = apiUser.loginUserStep(creds);
+        accessToken = loginResponse.as(UserLoginResponse.class).getAccessToken();
+
+        loginResponse.then()
+                .assertThat()
+                .statusCode(200)
+                .body( "accessToken", notNullValue());
+        System.out.println(loginResponse.body().asString());
+    }
 
     @Test
     @DisplayName("Авторизация с неверными данными")
     public void invalidLoginTest() {
         UserCreds creds = new UserCreds("fhfbdfh@mail.ru", "some-password");
         Response loginResponse = apiUser.loginUserStep(creds);
-        assertEquals(401, loginResponse.statusCode());
+
+        loginResponse.then()
+                .assertThat()
+                .statusCode(401)
+                .body("message", equalTo("email or password are incorrect"));
+        System.out.println(loginResponse.body().asString());
     }
 
     @AfterEach
     public void tearDown() {
-        apiUser.deleteUserStep(accessToken);
+        if (accessToken != null) {
+            apiUser.deleteUserStep(accessToken);
+        }
     }
 }
